fix repository generic type for client & commande

+++ src/main/java/licence/partiel20242025/repositories/LivreurRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface LivreurRepository extends JpaRepository<Livreur, Float>
+public interface LivreurRepository extends JpaRepository<Livreur, Integer>
 {
 
 }
